backend/requirements.txt
Authlib==1.5.2
blinker==1.9.0
base58==2.1.1
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
Flask==3.1.0
flask-cors==5.0.1
Flask-Session==0.8.0
flask-socketio==5.3.6
python-socketio==5.11.1
gunicorn==21.2.0
eventlet==0.36.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.19.0
psycopg2-binary==2.9.10
pycparser==2.22
PyNaCl==1.5.0
pytz==2025.2
redis==5.2.1
requests==2.32.3
schedule==1.2.2
urllib3==2.4.0
Werkzeug==3.1.3
